Extracted RegistrationIntent type
