+++ consul/consul_stub.go
@@ -105,10 +105,6 @@ func (c *ConsulStub) RegisterWithoutMarathonTaskTag(task *apps.Task, app *apps.A
 	c.services[service.ServiceId(serviceRegistration.ID)] = &serviceRegistration
 }
 
-func (c *ConsulStub) ServiceName(app *apps.App) string {
-	return c.consul.ServiceName(app)
-}
-
 func (c *ConsulStub) DeregisterByTask(taskId apps.TaskId) error {
 	if _, ok := c.failDeregisterByTaskForIds[taskId]; ok {
 		return fmt.Errorf("Consul stub programmed to fail when deregistering task of id %s", taskId.String())
