+++ service/service.go
@@ -40,5 +40,4 @@ type ServiceRegistry interface {
 	Register(task *apps.Task, app *apps.App) error
 	DeregisterByTask(taskId apps.TaskId) error
 	Deregister(toDeregister *Service) error
-	ServiceName(app *apps.App) string
 }
