+++ sync/sync_test.go
@@ -160,10 +160,6 @@ func (c *ConsulServicesMock) RegistrationsCount(instanceId string) int {
 	return c.registrations[instanceId]
 }
 
-func (c *ConsulServicesMock) ServiceName(app *apps.App) string {
-	return ""
-}
-
 func (c *ConsulServicesMock) DeregisterByTask(taskId apps.TaskId) error {
 	return nil
 }
